fix/Touch 오류
Touch가 한 손가락만 인식하는 버그 수정 및 BuildTest용 TMPro 코드

+++ Assets/Scripts/Contents/Note/Note/HoldNoteBody.cs
@@ -55,7 +55,7 @@ public class HoldNoteBody : Note
             {
                 if(lane >= holdNotes[curJudge].line && lane <= holdNotes[curJudge].line + holdNotes[curJudge].length)
                 {
-                    int tempJudge = holdNotes[curJudge].judge;
+                    float tempJudge = holdNotes[curJudge].judge;
                     if(curTime < tempJudge - (float)bpm / 600 * 16) 
                     {
                         return 0;
